pyproject.toml
[tool.poetry]
name = "basic-api"
version = "0.1.0"
description = ""
authors = ["mindahun21 <mndahun0@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.1.5"
djangorestframework = "^3.15.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
